video-analysis-model-main/requirements.txt
boto3
requests
redis
uuid
torch
openai-whisper
pandas
face_recognition
numpy<2
scipy
opencv-python
cmake
transformers
pyannote.audio
vocab
torchvision
moviepy
retrying
tiktoken
